Implement explicitly by Web3InterfaceAdapter
- Move InterfaceAdapter to ./types
- (also) Export all types in package root
- Remove Web3InterfaceAdapter's setProvider method (not part of iface)
- Define Web3InterfaceAdapterOptions explicitly, not related to shim
- Update Web3InterfaceAdapter class to implement InterfaceAdapter
