+++ packages/interface-adapter/lib/index.ts
@@ -1,2 +1,6 @@
 export { Web3Shim } from "./shim";
-export { createInterfaceAdapter, InterfaceAdapter } from "./adapter";
+export {
+  createInterfaceAdapter,
+  InterfaceAdapterOptions
+} from "./adapter";
+export * from "./adapter/types";
